+++ purchases/models.py
@@ -47,11 +47,11 @@ class Purchases(models.Model):
         for animal in animals:
             kg_totales.append(animal.cantidad * kg_cabeza)
 
-        animal_precio_kg =list(map(int,animals.values_list('precio_por_kg', flat=True)))
+        animal_precio_kg =list(map(float,animals.values_list('precio_por_kg', flat=True)))
 
         sub_totals= [a * b for a, b in zip(animal_precio_kg, kg_totales)]
 
-        animal_ivas = list(map(int,animals.values_list('iva', flat=True)))
+        animal_ivas = list(map(float,animals.values_list('iva', flat=True)))
 
         animal_totals = [a + b for a, b in zip(animal_ivas, sub_totals)]
 
@@ -68,11 +68,11 @@ class Purchases(models.Model):
 
         endorsed_checks = self.endorsed_checks
 
-        check_payed = sum([int(check.monto) for check in self_checks ])
+        check_payed = sum([float(check.monto) for check in self_checks ])
 
-        payments_payed = sum(list(map(int,payments.values_list('monto', flat=True))))
+        payments_payed = sum(list(map(float,payments.values_list('monto', flat=True))))
 
-        endorsed_payed = sum(list(map(int,endorsed_checks.values_list('monto', flat=True))))
+        endorsed_payed = sum(list(map(float,endorsed_checks.values_list('monto', flat=True))))
 
         total_payed = check_payed + payments_payed + endorsed_payed
