+++ sales/views.py
@@ -82,11 +82,11 @@ def sales_detail(request, id):
 
     animal_cabezas = list(map(int,sale_rows.values_list('cantidad', flat=True)))
 
-    animal_precio_kg =list(map(int,sale_rows.values_list('precio_por_kg', flat=True)))
+    animal_precio_kg =list(map(float,sale_rows.values_list('precio_por_kg', flat=True)))
 
     sub_totals= [a * b for a, b in zip(animal_precio_kg, kg_totales)]
 
-    animal_ivas = list(map(int,sale_rows.values_list('iva', flat=True)))
+    animal_ivas = list(map(float,sale_rows.values_list('iva', flat=True)))
 
     animal_totals = [a + b for a, b in zip(animal_ivas, sub_totals)]
 
@@ -106,8 +106,8 @@ def sales_detail(request, id):
 
     third_p_checks = sale.third_party_checks
 
-    total_payments_payed = sum(list(map(int,payments.values_list('monto', flat=True))))
-    total_checks_payed = sum(list(map(int,third_p_checks.values_list('monto', flat=True))))
+    total_payments_payed = sum(list(map(float,payments.values_list('monto', flat=True))))
+    total_checks_payed = sum(list(map(float,third_p_checks.values_list('monto', flat=True))))
 
     total_payed = total_payments_payed + total_checks_payed
